requirements.txt
Django==4.0.6
django-cors-headers==3.13.0
django-ignoretests==0.3.2
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf_yasg
psycopg2-binary==2.9.3
django_compressor==4.1
gunicorn==20.0.4
django_heroku
django-filter
django-environ
django-cors-headers==3.13.0
django-import-export==3.0.1
django-model-utils==4.3.1
pymupdf==1.21.0
pytesseract==0.3.10
pdf2image==1.16.0
pdfminer==20191125